% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axes.R
\name{ColAlpha}
\alias{ColAlpha}
\title{ColAlpha}
\usage{
ColAlpha(colvec, alpha = 0.5, maxColorValue = 255)
}
\arguments{
\item{maxColorValue}{}
}
\value{
Output of \code{rgb}.
}
\description{
Add alpha transparency to colors.
}

