% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirrorTicks.R
\name{GGMirrorTicks}
\alias{GGMirrorTicks}
\title{GGMirrorTicks}
\usage{
GGMirrorTicks(ggobj, sides = "bltr")
}
\arguments{
\item{ggobj}{output of \code{ggplot}}

\item{sides}{any of "bltr"}
}
\value{
\code{gtable} object to be drawn with \code{grid.draw}.
}
\description{
Adds mirrored tickmarks in ggplot panels. Especially when faceting where dup_axis() does not work (otherwise works well for single panel plots). Based on example provided by Sandy Muspratt on Stackoverflow: http://stackoverflow.com/a/29023682/143476.
}
