% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axes.R
\name{LogAxis}
\alias{LogAxis}
\title{LogAxis}
\usage{
LogAxis(side, labels = TRUE, minor = TRUE, decades.only = TRUE,
  major.labels = labels, minor.labels = FALSE, major.args = NULL,
  minor.args = NULL, ...)
}
\arguments{
\item{side}{1-4}

\item{labels}{TRUE/FALSE If FALSE, add minor tick labels like Igor Pro}

\item{minor}{TRUE/FALSE}

\item{decades.only}{TRUE/FALSE}

\item{major.labels}{TRUE/FALSE}

\item{minor.labels}{TRUE/FALSE}

\item{major.args}{list of tick arguments to \code{axis}}

\item{minor.args}{list of tick arguments to \code{axis}}

\item{...}{to be passed to \code{axis}}
}
\value{
List of tick positions and labels.
}
\description{
Logarithmic axis
}

