% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axes.R
\name{MTitle}
\alias{MTitle}
\title{MTitle}
\usage{
MTitle(main = NULL, xlab = NULL, ylab = NULL, ylab2 = NULL,
  line = NULL, line.frac = 2/5, ...)
}
\arguments{
\item{line.frac}{}
}
\value{
\code{NULL}
}
\description{
Like \code{title} but places labels in outer margin.
}

