% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Recsub}
\alias{Recsub}
\title{Recsub}
\usage{
Recsub(string, replace, ...)
}
\arguments{
\item{string}{character string to replace}

\item{replace}{list of string pairs to be passed to \code{gsub}}
}
\value{
Character string, same length as \code{string}.
}
\description{
Recursive string substitution.
}

