% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{StripWhite}
\alias{StripWhite}
\title{StripWhite}
\usage{
StripWhite(x)
}
\arguments{
\item{x}{character string}
}
\value{
Character string stripped of white characters.
}
\description{
Strip white characters.
}
