% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axes.R
\name{Subsetinrange}
\alias{Subsetinrange}
\title{Subsetinrange}
\usage{
Subsetinrange(x, lim)
}
\arguments{
\item{x}{numeric vector}

\item{lim}{interval limits}
}
\value{
Selected values.
}
\description{
Helper function. Returns values falling within a predetermined interval.
}

