% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axes.R
\name{TextLabel}
\alias{TextLabel}
\title{TextLabel}
\usage{
TextLabel(location, ..., yoffset = 0, expansion = 0.1)
}
\arguments{
\item{location}{one of "topleft", "topright", "bottomleft", "bottomright"}

\item{...}{to \code{text}}

\item{yoffset}{generally integer values to be multipliedy by \code{par("cxy")}}

\item{expansion}{}
}
\value{
\code{NULL}
}
\description{
Convenience functions for adding letters and labels to figures. Adds padding not provided by \code{text(par("usr")[1], par("usr")[4], ...)}.
}

